[UPD] A function which will add a large block to the large memory zone, and then return this block to the malloc has been implemented.
