+++ includes/ft_malloc.h
@@ -34,4 +34,9 @@ typedef struct s_memory
 
 extern t_memory glob_memory;
 
+t_block *find_block(size_t size);
+t_block *split_block(t_block *current, size_t size);
+void extend_heap(t_mem_group *mem_group, size_t size);
+
+
 #endif
